% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_yo_id}
\alias{add_yo_id}
\title{Add yo ID to gliderdf}
\usage{
add_yo_id(df)
}
\arguments{
\item{df}{input dataframe that must contain a column named "cast"
with values "Downcast" and "Upcast"}
}
\value{
appended dataframe with new "yo_id" column
}
\description{
Sequentially count downcast and upcast cycles
}
