% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{depthInt}
\alias{depthInt}
\title{Interpolate depth over time}
\usage{
depthInt(inGliderdf, CTD = TRUE)
}
\arguments{
\item{inGliderdf}{input dataframe that must contain a depth measure}

\item{CTD}{boolean, use CTD-derived depth, else use m_depth}
}
\value{
outputted dataframe with time and depth
}
\description{
Uses zoo package to build dataframe with interpolated
depth values through all available times
}
