% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gliderGPS_to_dd}
\alias{gliderGPS_to_dd}
\title{Convert glider GPS to decimal degrees}
\usage{
gliderGPS_to_dd(gliderGPS)
}
\arguments{
\item{gliderGPS}{input value to be converted}
}
\value{
numeric output value
}
\description{
Helper function that converts m_gps_* and
m_lat/m_lon to decimal degrees.
}
