% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gotoLoad}
\alias{gotoLoad}
\title{Parse goto_l*.ma files}
\usage{
gotoLoad(inFile)
}
\arguments{
\item{inFile}{user-specified path to file}
}
\value{
dataframe of waypoints and associated metadata
}
\description{
Dynamically parse glider goto files
and pull out useful variables.
}
