% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identify_casts_smooth}
\alias{identify_casts_smooth}
\title{Track glider state through time}
\usage{
identify_casts_smooth(data, surface_threshold, rolling_window_size)
}
\arguments{
\item{data}{input dataframe that must contain a depth column}

\item{surface_threshold}{numeric, how deep is considered "surface"?}

\item{rolling_window_size}{numeric, how many depth samples to check across?}
}
\value{
appended dataframe with new "cast" column indicating state
}
\description{
Using glider depth, determine if the vehicle is ascending, descending,
or at the surface and label it as such.
}
