% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osgUtils-package.R
\docType{package}
\name{osgUtils-package}
\alias{osgUtils}
\alias{osgUtils-package}
\title{osgUtils: Slocum Glider Data Wrangling}
\description{
Tools to help process Slocum glider data.
}
\author{
\strong{Maintainer}: Garrett Miller \email{millerg1@usf.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
