% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{ssv_to_df}
\alias{ssv_to_df}
\title{Convert ssv generated by dbd2asc into dataframe}
\usage{
ssv_to_df(inputFile, headerSize = 14)
}
\arguments{
\item{inputFile}{user-specified path to file}

\item{headerSize}{number of lines to consider as "header"}
}
\value{
dataframe of ssv minus the header
}
\description{
Builds dataframe from standard ssv generated
by dbd2asc on any format of glider data
}
